Fix mint flow
