+++ src/context/ViewContext.jsx
@@ -8,7 +8,7 @@ export const ViewStates = {
   CREATOR: "CREATOR",
   MINT_LOADING: "MINT_LOADING",
   MINT: "MINT",
-  MINT: "MINT_CONFIRM",
+  MINT_CONFIRM: "MINT_CONFIRM",
   MINT_COMPLETE: "MINT_COMPLETE"
 }
